get_data/requirements.txt
agents==1.4.0
numpy==2.3.2
openai==1.99.9
pandas==2.3.1
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
